+++ .gitignore
@@ -6,3 +6,4 @@ yarn-error.log
 **/*.rs.bk
 .env.*
 !.env-sample
+.DS_Store
