+++ src/types/responses/At.ts
@@ -1,0 +1,6 @@
+import { BlockHash } from '@polkadot/types/interfaces';
+
+export interface IAt {
+	hash: string | BlockHash;
+	height: string;
+}
