+++ src/types/responses/SanitizedArgs.ts
@@ -1,0 +1,7 @@
+import { ISanitizedCall } from '.';
+
+export interface ISanitizedArgs {
+	call?: ISanitizedCall;
+	calls?: ISanitizedCall[];
+	[key: string]: unknown;
+}
