+++ src/types/responses/SanitizedCall.ts
@@ -1,0 +1,8 @@
+import { ISanitizedArgs } from '.';
+
+export interface ISanitizedCall {
+	[key: string]: unknown;
+	method: string;
+	callIndex: Uint8Array | string;
+	args: ISanitizedArgs;
+}
