+++ src/types/responses/SanitizedEvent.ts
@@ -1,0 +1,6 @@
+import { EventData } from '@polkadot/types/generic/Event';
+
+export interface ISanitizedEvent {
+	method: string;
+	data: EventData;
+}
